feat: add sign up page

+++ Makefile
@@ -1,5 +1,5 @@
 clean-build:
-	rd /s /q "build"
+	if exist "build" rd /s /q "build"
 
 run-chrome: clean-build
 	flutter run -d chrome --dart-define-from-file=.env
